+++ NEWS.md
@@ -1,4 +1,4 @@
-## 2.0.2 ##
+## 2.1.0 ##
 * Add support for GitLab CI (#190, @enbrown).
 * Update exclusion documentation to include line_exclusions and function
   exclusions (#191).